Clean up of tests
