+++ tests/functional_tests/ffront_tests/test_execution.py
@@ -20,7 +20,7 @@ from typing import TypeVar
 import numpy as np
 import pytest
 
-from functional.common import DimensionKind, GTTypeError
+from functional.common import DimensionKind
 from functional.fencil_processors.runners import gtfn_cpu, roundtrip
 from functional.ffront.decorator import field_operator, program, scan_operator
 from functional.ffront.fbuiltins import (
@@ -1108,39 +1108,3 @@ def test_domain_tuple(fieldview_backend):
 
     np.allclose(np.asarray(a), a)
     np.allclose(expected, b)
-
-
-def test_domain_slices(fieldview_backend):
-    size = 10
-    a = np_as_located_field(IDim)(np.ones((size,)))
-    b = np_as_located_field(IDim)(np.ones((size,)))
-
-    @field_operator(backend=fieldview_backend)
-    def fieldop_domain_tuple(a: Field[[IDim], float64]) -> Field[[IDim], float64]:
-        return a
-
-    @program
-    def program_domain_tuple(a: Field[[IDim], float64], b: Field[[IDim], float64]):
-        fieldop_domain_tuple(a, out=b[0:1], domain={IDim: (1, 9)})
-
-    program_domain_tuple(a, b, offset_provider={})
-
-
-def test_domain_exception_5():
-    size = 10
-    a = np_as_located_field(IDim)(np.ones((size,)))
-    out_field = np_as_located_field(IDim)(np.ones((size,)))
-
-    @field_operator
-    def domain_and_slicing_fieldop(a: Field[[IDim], float64]) -> Field[[IDim], float64]:
-        return a
-
-    @program
-    def domain_and_slicing_program(a: Field[[IDim], float64], out_field: Field[[IDim], float64]):
-        domain_and_slicing_fieldop(a, out=out_field[0:1], domain={IDim: (0, 1)})
-
-    with pytest.raises(
-        GTTypeError,
-        match=(r"Either only domain or slicing allowed"),
-    ):
-        domain_and_slicing_program(a, out_field, offset_provider={})
